+++ src/platform/ui_engine/patterns/index.tsx
@@ -1,6 +1,7 @@
 export * from "./actions";
 export * from "./combobox";
 export * from "./creatable-search";
+export * from "./creatable-multi-select";
 export * from "./hooks";
 export * from "./inline-edit";
 export * from "./simple-text-editor";